Fixed get-daily-notes.ts and printToConsole()
Also made it so that printToConsole() can accept any value as input.

Also made getPriorNotes() and the On this day view be based on the beginning of the day (12:00:00.000 AM) instead of the current time.

Also made function `getModifiedFolderAndFormat()` to standardize the inputs of the different functions in get-daily-notes.ts.