Update onClick() in note-previews
Made it so that if you click a note preview, instead of just replacing the first markdown view, it first checks to see if any of them are the same TFile. If they're not, it checks to see if any of them are a daily note. If none of them are, THEN it replaces the first TFile.